+++ crud/upsert.lua
@@ -44,10 +44,10 @@ local function upsert_on_storage(space_name, tuple, operations, opts)
     -- add_space_schema_hash is true only in case of upsert_object
     -- the only one case when reloading schema can avoid insert error
     -- is flattening object on router
-    return schema.wrap_box_space_func_result(space, 'upsert', {tuple, operations}, {
+    return schema.wrap_func_result(space, space.upsert, {
         add_space_schema_hash = opts.add_space_schema_hash,
         fetch_latest_metadata = opts.fetch_latest_metadata,
-    })
+    }, space, tuple, operations)
 end
 
 upsert.storage_api = {[UPSERT_FUNC_NAME] = upsert_on_storage}