+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+* drop wrap_box_space_func_result to cut allocations and speed up storage calls.
+
 ## [1.6.1] - 19-09-25
 
 ### Added